+++ RNG_Test/Tests/BYTE_Testing/Unsigned_BYTE/Unsigned_BYTE_Incorrect_Order.cs
@@ -19,7 +19,7 @@ namespace RNG_Test
 		/*
 		===============================================================================================
 		PURPOSE:
-		Test the Random Number Generator for and Unsigned BYTE when the parameters are provided in the
+		Test the Random Number Generator for an Unsigned BYTE when the parameters are provided in the
 		incorrect order => Maximum value, Minimum value
 		-----------------------------------------------------------------------------------------------
 		OUTPUT: