+++ RNG_Test/Tests/Integer_Testing/Unsigned_Integer/Unsigned_Integer_Correct_Order.cs
@@ -12,17 +12,17 @@ using Con = System.Console;
 
 namespace RNG_Test
 {
-	partial class BYTE_Testing
+	partial class Integer_Testing
 	{
-		private void Signed_BYTE_Correct_Order()
+		private void Unsigned_Integer_Correct_Order()
 		/*
 		===============================================================================================
 		PURPOSE:
-		Test the Random Number Generator for a Signed BYTE when the parameters are provided in the
+		Test the Random Number Generator for an Unsigned Integer when the parameters are provided in the
 		correct order => Minimum value, Maximum value
 		-----------------------------------------------------------------------------------------------
 		OUTPUT:
-		Randomly generated Signed BYTE values.
+		Randomly generated Signed Integer values.
 		===============================================================================================
 		*/
 		{
@@ -30,8 +30,8 @@ namespace RNG_Test
 			// Variables - Standard
 			//=============
 			RNG MyRandomNumber = new RNG();
-			sbyte MinValue = sbyte.MinValue;
-			sbyte MaxValue = sbyte.MaxValue;
+			uint MinValue = uint.MinValue;
+			uint MaxValue = uint.MaxValue;
 
 			//=============
 			// Body
@@ -43,6 +43,6 @@ namespace RNG_Test
 			Con.Write("{0}", MyRandomNumber.Generate(MinValue, MaxValue) + "\t");
 			Con.Write("{0}", MyRandomNumber.Generate(MinValue, MaxValue) + "\t");
 			Con.WriteLine("{0}", MyRandomNumber.Generate(MinValue, MaxValue));
-		} // private void Signed_BYTE_Correct_Order
-	} // partial class BYTE_Testing
+		} // private void Signed_Integer_Correct_Order
+	} // partial class Integer_Testing
 } // namespace RNG_Test