+++ RNG_Test/Tests/Integer_Testing/Unsigned_Integer/Unsigned_Integer_Loop.cs
@@ -14,12 +14,11 @@ namespace RNG_Test
 {
 	partial class Integer_Testing
 	{
-		private void Signed_Integer_Correct_Order()
+		private void Unsigned_Integer_Loop()
 		/*
 		===============================================================================================
 		PURPOSE:
-		Test the Random Number Generator for a Signed Integer when the parameters are provided in the
-		correct order => Minimum value, Maximum value
+		Test the Random Number Generator for a Signed Integer within a loop.
 		-----------------------------------------------------------------------------------------------
 		OUTPUT:
 		Randomly generated Signed Integer values.
@@ -30,19 +29,22 @@ namespace RNG_Test
 			// Variables - Standard
 			//=============
 			RNG MyRandomNumber = new RNG();
-			int MinValue = int.MinValue;
-			int MaxValue = int.MaxValue;
+			uint MinValue = uint.MinValue;
+			uint MaxValue = uint.MaxValue;
 
 			//=============
 			// Body
 			//=============
 			Con.WriteLine();
-			Con.WriteLine(@"Pass the valid values in the correct order (MinValue, MaxValue)");
-			Con.Write("{0}", MyRandomNumber.Generate(MinValue, MaxValue) + "\t");
-			Con.Write("{0}", MyRandomNumber.Generate(MinValue, MaxValue) + "\t");
-			Con.Write("{0}", MyRandomNumber.Generate(MinValue, MaxValue) + "\t");
-			Con.Write("{0}", MyRandomNumber.Generate(MinValue, MaxValue) + "\t");
-			Con.WriteLine("{0}", MyRandomNumber.Generate(MinValue, MaxValue));
-		} // private void Signed_Integer_Correct_Order
+			Con.WriteLine(@"Generate 5 random numbers 3 times");
+			for (int i = 0; i < 3; i++)
+			{
+				Con.Write("{0}", MyRandomNumber.Generate(MinValue, MaxValue) + "\t");
+				Con.Write("{0}", MyRandomNumber.Generate(MinValue, MaxValue) + "\t");
+				Con.Write("{0}", MyRandomNumber.Generate(MinValue, MaxValue) + "\t");
+				Con.Write("{0}", MyRandomNumber.Generate(MinValue, MaxValue) + "\t");
+				Con.WriteLine("{0}", MyRandomNumber.Generate(MinValue, MaxValue));
+			}
+		} // private void Unsigned_Integer_Loop
 	} // partial class Integer_Testing
 } // namespace RNG_Test