+++ RNG_Test/Tests/Integer_Testing/Unsigned_Integer/Unsigned_Integer_Values_Outside_Range.cs
@@ -12,22 +12,22 @@ using Con = System.Console;
 
 namespace RNG_Test
 {
-	partial class BYTE_Testing
+	partial class Integer_Testing
 	{
-		private void Unsigned_BYTE_Values_Outside_Range()
+		private void Unsigned_Integer_Values_Outside_Range()
 		/*
 		===============================================================================================
 		PURPOSE:
-		Test the Random Number Generator for an Unsigned BYTE when the parameters that are provided
-		are outside the range of an Unsigned BYTE.
+		Test the Random Number Generator for and Signed Integer when the parameters that are provided
+		are outside the range of an Unsigned Integer.
 		-----------------------------------------------------------------------------------------------
 		OUTPUT:
-		An error message if the values are outside the range; otherwise a randomly generated Unsigned
-		BYTE value.
+		An error message if the values are outside the range; otherwise a randomly generated Signed
+		Integer value.
 		-----------------------------------------------------------------------------------------------
 		NOTES:
 		- When a value is passed to the RNG, the appropriate overloaded function will be called.
-		  Because of this, a value outside of a larger unit becomes an int/long depending on the
+		  Because of this, a value outside of an Unsigned Integer becomes a larger unit depending on the
 		  value.  To ensure the RESULTS are not outside the range, the calling process needs to
 		  validate the values PRIOR to calling.
 		===============================================================================================
@@ -37,15 +37,22 @@ namespace RNG_Test
 			// Variables - Standard
 			//=============
 			RNG MyRandomNumber = new RNG();
-			byte MinValue = byte.MinValue;
-			byte MaxValue = byte.MaxValue;
+			uint MinValue = uint.MinValue;
+			uint MaxValue = uint.MaxValue;
 			bool IsValid = true;  // Assume the value is valid
 
 			//=============
-			// Variables - outside the scope of BYTE
+			// Variables - outside the scope of an Unsigned Integer
 			//=============
-			int Bad_MinValue = byte.MinValue - 1;
-			int Bad_MaxValue = byte.MaxValue + 1;
+			long Bad_MinValue = uint.MinValue;
+			long Bad_MaxValue = uint.MaxValue;
+
+			//=============
+			// Setup Environment
+			//=============
+			// Performing this calculation at declaration causes an overflow error.
+			Bad_MinValue -= 1;
+			Bad_MaxValue += 1;
 
 			//=============
 			// Body
@@ -54,16 +61,16 @@ namespace RNG_Test
 			Con.WriteLine(@"Pass values outside the range");
 
 			// Validate the minimum value
-			if (Bad_MinValue < byte.MinValue)
+			if (Bad_MinValue < uint.MinValue)
 			{
-				Con.WriteLine(@"Intended value ({0}) cannot be converted to UNSIGNED BYTE.", Bad_MinValue);
+				Con.WriteLine(@"Intended value ({0}) cannot be converted to UNSIGNED Integer.", Bad_MinValue);
 				IsValid = false;
 			}
 
 			// Validate the maximum value
-			if (Bad_MaxValue > byte.MaxValue)
+			if (Bad_MaxValue > uint.MaxValue)
 			{
-				Con.WriteLine(@"Intended value ({0}) cannot be converted to UNSIGNED BYTE.", Bad_MaxValue);
+				Con.WriteLine(@"Intended value ({0}) cannot be converted to UNSIGNED Integer.", Bad_MaxValue);
 				IsValid = false;
 			}
 
@@ -76,6 +83,6 @@ namespace RNG_Test
 				Con.Write("{0}", MyRandomNumber.Generate(MinValue, Bad_MaxValue) + "\t");
 				Con.WriteLine("{0}", MyRandomNumber.Generate(Bad_MinValue, Bad_MaxValue));
 			}
-		} // private void Unsigned_BYTE_Values_Outside_Range
-	} // partial class BYTE_Testing
+		} // private void Unsigned_Integer_Values_Outside_Range
+	} // partial class Integer_Testing
 } // namespace RNG_Test